package 명에 대문자 변경 및 PullRequest 피브백 반영
