--- src/main/java/com/market/Server/advice/exception/CUserNotFoundException.java
@@ -1,15 +1,0 @@
-package com.market.Server.advice.exception;
-
-public class CUserNotFoundException extends RuntimeException {
-    public CUserNotFoundException(String msg, Throwable t) {
-        super(msg, t);
-    }
-
-    public CUserNotFoundException(String msg) {
-        super(msg);
-    }
-
-    public CUserNotFoundException() {
-        super();
-    }
-}
