--- src/main/java/com/market/Server/advice/exception/DuplicateIdException.java
@@ -1,9 +1,0 @@
-package com.market.Server.advice.exception;
-
-public class DuplicateIdException extends RuntimeException {
-  
-  public DuplicateIdException(String msg) {
-    super(msg);
-  }
-  
-}
