--- src/main/java/com/market/Server/config/DatabaseConfig.java
@@ -1,40 +1,0 @@
-package com.market.Server.config;
-
-import org.apache.ibatis.session.SqlSessionFactory;
-import org.mybatis.spring.SqlSessionFactoryBean;
-import org.mybatis.spring.SqlSessionTemplate;
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-import javax.sql.DataSource;
-
-@Configuration
-@MapperScan(basePackages = "com.market.Server")
-@EnableTransactionManagement // DataSourceTransactionManager Bean을 Transaction Manager로 사용
-public class DatabaseConfig {
-
-  @Bean
-  public SqlSessionFactory sqlSessionFactory(DataSource dataSource) throws Exception {
-    final SqlSessionFactoryBean sessionFactory = new SqlSessionFactoryBean();
-    sessionFactory.setDataSource(dataSource);
-    // TypeAlias로 설정할 클래스들이 있는 패키지를 설정하면 DTO에 @Alias("aliasName")으로 typeAlias를 설정 가능
-    sessionFactory.setTypeAliasesPackage("com.market.Server.dto.");
-    PathMatchingResourcePatternResolver resolver = new PathMatchingResourcePatternResolver();
-    sessionFactory.setMapperLocations(resolver.getResources("classpath:mybatis/mapper/*.xml"));
-    return sessionFactory.getObject();
-}
-
-  @Bean
-  public SqlSessionTemplate sqlSessionTemplate(SqlSessionFactory sqlSessionFactory)
-      throws Exception {
-    final SqlSessionTemplate sqlSessionTemplate = new SqlSessionTemplate(sqlSessionFactory);
-    return sqlSessionTemplate;
-  }
-
-
-}
-
-
