--- src/main/java/com/market/Server/model/response/ListResult.java
@@ -1,12 +1,0 @@
-package com.market.Server.model.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-public class ListResult<T> extends CommonResult {
-    private List<T> list;
-}
