--- src/main/java/com/market/Server/model/response/SingleResult.java
@@ -1,10 +1,0 @@
-package com.market.Server.model.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SingleResult<T> extends CommonResult {
-    private T data;
-}
