--- src/main/java/com/market/Server/service/UserService.java
@@ -1,10 +1,0 @@
-package com.market.Server.service;
-
-import com.market.Server.dto.UserDTO;
-
-public interface UserService {
-
-    // 회원 가입 처리
-    void register(UserDTO userProfile);
-
-}
