+++ src/main/java/com/market/server/ServerApplication.java
@@ -1,0 +1,15 @@
+package com.market.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
+
+@SpringBootApplication
+@EnableCaching // Spring에서 Caching을 사용하겠다고 선언한다.
+public class ServerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ServerApplication.class, args);
+    }
+
+}
