+++ src/main/java/com/market/server/advice/exception/CUserNotFoundException.java
@@ -1,0 +1,15 @@
+package com.market.server.advice.exception;
+
+public class CUserNotFoundException extends RuntimeException {
+    public CUserNotFoundException(String msg, Throwable t) {
+        super(msg, t);
+    }
+
+    public CUserNotFoundException(String msg) {
+        super(msg);
+    }
+
+    public CUserNotFoundException() {
+        super();
+    }
+}
