+++ src/main/java/com/market/server/config/SwaggerConfiguration.java
@@ -1,0 +1,30 @@
+package com.market.server.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@EnableSwagger2
+public class SwaggerConfiguration {
+    @Bean
+    public Docket swaggerApi() {
+        return new Docket(DocumentationType.SWAGGER_2).apiInfo(swaggerInfo()).select()
+                .apis(RequestHandlerSelectors.basePackage("com.market.server.controller"))
+                .paths(PathSelectors.any())
+                .build()
+                .useDefaultResponseMessages(false); // 기본으로 세팅되는 200,401,403,404 메시지를 표시 하지 않음
+    }
+
+    private ApiInfo swaggerInfo() {
+        return new ApiInfoBuilder().title("Spring API Documentation")
+                .description("앱 개발시 사용되는 서버 API에 대한 연동 문서입니다")
+                .license("junshock5").licenseUrl("https://github.com/f-lab-edu/used-market-server/tree/feature/2/Login").version("1").build();
+    }
+}
