+++ src/main/java/com/market/server/model/response/ListResult.java
@@ -1,0 +1,12 @@
+package com.market.server.model.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+public class ListResult<T> extends CommonResult {
+    private List<T> list;
+}
