+++ src/main/java/com/market/server/model/response/SingleResult.java
@@ -1,0 +1,10 @@
+package com.market.server.model.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SingleResult<T> extends CommonResult {
+    private T data;
+}
