+++ src/main/java/com/market/server/service/ResponseService.java
@@ -1,0 +1,74 @@
+package com.market.server.service;
+
+import com.market.server.model.response.CommonResult;
+import com.market.server.model.response.ListResult;
+import com.market.server.model.response.SingleResult;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+
+// 해당 Class가 서비스(서버에서 클라이언트에게 응답할때 response 패킷을 정의) 레이어 클래스라는 것을 spring framework에 정의한다.
+// response를 class로 분리한 이유는 해당 클래스를 사용하는 곳에서 ExceptionHandler의 형태를 직접 정의한다.
+@Service
+public class ResponseService {
+    // enum으로 api 요청 결과에 대한 code, message를 정의합니다.
+    public enum CommonResponse {
+        SUCCESS(0, "성공하였습니디."),
+        FAIL(-1, "실패하였습니다.");
+        int code;
+        String msg;
+
+        CommonResponse(int code, String msg) {
+            this.code = code;
+            this.msg = msg;
+        }
+
+        public int getCode() {
+            return code;
+        }
+
+        public String getMsg() {
+            return msg;
+        }
+    }
+
+    // 단일건 결과를 처리하는 메소드
+    public <T> SingleResult<T> getSingleResult(T data) {
+        SingleResult<T> result = new SingleResult<>();
+        result.setData(data);
+        setSuccessResult(result);
+        return result;
+    }
+
+    // 다중건 결과를 처리하는 메소드
+    public <T> ListResult<T> getListResult(List<T> list) {
+        ListResult<T> result = new ListResult<>();
+        result.setList(list);
+        setSuccessResult(result);
+        return result;
+    }
+
+    // 성공 결과만 처리하는 메소드
+    public CommonResult getSuccessResult() {
+        CommonResult result = new CommonResult();
+        setSuccessResult(result);
+        return result;
+    }
+
+    // 실패 결과만 처리하는 메소드
+    public CommonResult getFailResult(int code, String msg) {
+        CommonResult result = new CommonResult();
+        result.setSuccess(false);
+        result.setCode(code);
+        result.setMsg(msg);
+        return result;
+    }
+
+    // 결과 모델에 api 요청 성공 데이터를 세팅해주는 메소드
+    private void setSuccessResult(CommonResult result) {
+        result.setSuccess(true);
+        result.setCode(CommonResponse.SUCCESS.getCode());
+        result.setMsg(CommonResponse.SUCCESS.getMsg());
+    }
+}
