+++ src/main/resources/application.properties
@@ -6,3 +6,9 @@ spring.datasource.password=wnstjr1026
 # message
 spring.messages.basename=i18n/exception
 spring.messages.encoding=UTF-8
+
+# redis
+spring.cache.type=redis
+spring.redis.host=localhost
+spring.redis.port=6379
+spring.redis.password=
