+++ src/main/resources/mybatis/mapper/userMapper.xml
@@ -3,29 +3,29 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.market.Server.mapper.UserProfileMapper">
+<mapper namespace="com.market.server.mapper.UserProfileMapper">
 
-    <insert id="register" parameterType="com.market.Server.dto.UserDTO">
+    <insert id="register" parameterType="com.market.server.dto.UserDTO">
         INSERT INTO userprofile (
-            id
-            , pw
-            , name
-            , phone
-            , address
+        id
+        , pw
+        , name
+        , phone
+        , address
 
         ) VALUES (
-            #{id}
-            , #{pw}
-            , #{name}
-            , #{phone}
-            , #{address}
+        #{id}
+        , #{pw}
+        , #{name}
+        , #{phone}
+        , #{address}
         )
     </insert>
 
-    <select id="findByIdAndPassword" resultType="com.market.Server.dto.UserDTO">
+    <select id="findByIdAndPassword" resultType="com.market.server.dto.UserDTO">
         SELECT *
         FROM userprofile
-        WHERE  id = #{id}
+        WHERE id = #{id}
         AND pw = #{pw}
         AND status != 'DELETE'
     </select>
@@ -35,5 +35,14 @@
         FROM userprofile
         WHERE id = #{id}
     </select>
+
+<!--    <select id="getUserProfile" parameterType="com.market.server.dto.UserDTO">-->
+<!--        SELECT id,pw,name,phone,address,status-->
+<!--        FROM UserProfile-->
+<!--        WHERE id= #{id}-->
+<!--    </select>-->
+
+
+
 
 </mapper>
