+++ src/test/java/com/market/server/ServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.market.server;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
